[rb] implement Remote::Driver as a class not a module
