+++ rb/lib/selenium/webdriver/remote/driver.rb
@@ -26,16 +26,23 @@ module Selenium
       # @api private
       #
 
-      module Driver
+      class Driver < WebDriver::Driver
+        include DriverExtensions::UploadsFiles
+        include DriverExtensions::TakesScreenshot
+        include DriverExtensions::HasSessionId
+        include DriverExtensions::Rotatable
+        include DriverExtensions::HasRemoteStatus
+        include DriverExtensions::HasWebStorage
 
-        def self.new(**opts)
+        def initialize(opts = {})
           listener = opts.delete(:listener)
-          bridge = Bridge.handshake(opts)
-          if bridge.dialect == :w3c
-            W3C::Driver.new(bridge, listener: listener)
-          else
-            OSS::Driver.new(bridge, listener: listener)
+          @bridge = Bridge.handshake(opts)
+          if @bridge.dialect == :oss
+            extend DriverExtensions::HasTouchScreen
+            extend DriverExtensions::HasLocation
+            extend DriverExtensions::HasNetworkConnection
           end
+          super(@bridge, listener: listener)
         end
 
       end # Driver
